:rocket: update: change models to underscored

+++ src/controllers/vote.controller.js
@@ -8,12 +8,12 @@ const createVote = async (req, res) => {
   try {
     const post = await Post.findOne({ where: { id: postId } })
     if (post) {
-      const existVote = await Vote.findOne({ where: { postId, userId } })
+      const existVote = await Vote.findOne({ where: { post_id: postId, user_id: userId } })
       if (dir === 1) {
         if (existVote) {
           return res.status(409).json({ message: 'Vote already exits' })
         }
-        await Vote.create({ postId, userId })
+        await Vote.create({ post_id: postId, user_id: userId })
         return res.status(201).json({
           message: 'Succefully added vote'
         })
@@ -21,7 +21,7 @@ const createVote = async (req, res) => {
         if (!existVote) {
           return res.status(404).json({ message: 'Vote not exits' })
         }
-        await Vote.destroy({ where: { postId, userId } })
+        await Vote.destroy({ where: { post_id: postId, user_id: userId } })
         return res.status(204).json({
           message: 'Succefully deleted vote'
         })
